# src/main/java/fsa/fresher/pos/api/dto/UserUpdateDto.java
package fsa.fresher.pos.api.dto;

import fsa.fresher.pos.user.UserRole;
import fsa.fresher.pos.user.UserStatus;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.Size;

public class UserUpdateDto {
    private String name;
    @Email
    private String email;
    @Size(min = 8)
    private String password;
    private UserRole role;
    private UserStatus status;

    public String getName() { return name; }
    public void setName(String name) { this.name = name; }
    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }
    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }
    public UserRole getRole() { return role; }
    public void setRole(UserRole role) { this.role = role; }
    public UserStatus getStatus() { return status; }
    public void setStatus(UserStatus status) { this.status = status; }
}